# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokümantasyonu",
        "description": "Bu API, sipariş oluşturma ve yönetme işlemleri için kullanılır.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9002",
            "description": "Local Geliştirme Sunucusu"
        }
    ],
    "paths": {
...
        "/api/order": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get all orders",
                "description": "Fetches a list of all orders.",
                "operationId": "getAllOrders",
                "responses": {
                    "200": {
                        "description": "List of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "customer_id": {
                                                "type": "integer",
                                                "example": 123
                                            },
                                            "items": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "productId": {
                                                            "type": "integer",
                                                            "example": 100
                                                        },
                                                        "quantity": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "unitPrice": {
                                                            "type": "string",
                                                            "example": "10.50"
                                                        },
                                                        "total": {
                                                            "type": "string",
                                                            "example": "21.00"
                                                        },
                                                        "discountAmount": {
                                                            "type": "string",
                                                            "example": "2.00"
                                                        },
                                                        "discountedTotal": {
                                                            "type": "string",
                                                            "example": "19.00"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "total": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 250.5
                                            },
                                            "totalDiscount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 0
                                            },
                                            "discountedTotal": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 0
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Orders not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Siparişler bulunamadı."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Create an order",
                "description": "Create a new order and return the created order.",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "items"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Siparişiniz başarıyla oluşturulmuştur."
                                        },
                                        "order_id": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "total": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 120
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid request data."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order/{id}": {
            "delete": {
                "tags": [
                    "Order"
                ],
                "summary": "Delete an order",
                "description": "Deletes an existing order by ID.",
                "operationId": "deleteOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sipariş silindi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sipariş Bulunamadı"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}